// src/types/Row.ts
export interface Row {
  created_dt: string
  data_source_modified_dt: string
  entity_type: string
  operating_status: string | null
  legal_name: string | null
  dba_name: string | null
  physical_address: string | null
  p_street: string | null
  p_city: string | null
  p_state: string | null
  p_zip_code: string | null
  phone: string | null
  mailing_address: string | null
  m_street: string | null
  m_city: string | null
  m_state: string | null
  m_zip_code: string | null
  usdot_number: string | null
  mc_mx_ff_number: string | null
  power_units: string | null
  mcs_150_form_date: string | null
  out_of_service_date: string | null
  state_carrier_id_number: string | null
  duns_number: string | null
  drivers: string | null
  mcs_150_mileage_year: string | null
  credit_score: string | null
  record_status: string | null
}